requirements.txt
pdf2image
pytesseract
numpy
pdfminer.six
chardet
matplotlib